Revert "Merge remote-tracking branch 'origin/feature/RealDragon' into feature/map"
This reverts commit d5688947cb915481686f560c0b0e0ea5ee0a419e, reversing
changes made to 457e09300769d09926e66b1b23a7e7ae35631c40.

+++ Source/RuinLooters/SaveGame/RLSaveGame.h
@@ -44,10 +44,10 @@ public:
 	int32 PlayerMoney;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Player")
-	float CurrentHp;
+	int32 CurrentHp;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Player")
-	float MaxHp;
+	int32 MaxHp;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Player")
 	int32 AttackDamage;
